refactor: dedicated file for plugin settings code
